test: TECH-139: Code review fixes and classes renaming.
