+++ src/main/java/org/hisp/dhis/Main.java
@@ -15,8 +15,8 @@ import org.hisp.dhis.cache.EntitiesCache;
 import org.hisp.dhis.locust.LocustConfig;
 import org.hisp.dhis.locust.LocustSlave;
 import org.hisp.dhis.tasks.AddTeiTask;
+import org.hisp.dhis.tasks.GetHeavyAnalyticsRandomTask;
 import org.hisp.dhis.tasks.GetHeavyAnalyticsTask;
-import org.hisp.dhis.tasks.GetHeavyAnalyticsTaskNoRandom;
 import org.hisp.dhis.tasks.LoginTask;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -71,8 +71,8 @@ public class Main
 
         locust.run(
                 new AddTeiTask( 50, cache ),
-                new GetHeavyAnalyticsTaskNoRandom( 30 ),
-                new GetHeavyAnalyticsTask( 30, cache )
+                new GetHeavyAnalyticsTask( 30, cfg.analyticsApiVersion() ),
+                new GetHeavyAnalyticsRandomTask( 30, cfg.analyticsApiVersion(), cache )
         );
     }
 }
