+++ src/main/java/org/hisp/dhis/tasks/GetHeavyAnalyticsRandomTask.java
@@ -2,11 +2,16 @@ package org.hisp.dhis.tasks;
 
 import static com.google.api.client.http.HttpStatusCodes.STATUS_CODE_OK;
 import static io.restassured.RestAssured.given;
+import static java.lang.String.format;
 import static java.lang.String.join;
 import static java.lang.String.valueOf;
 import static java.util.Arrays.asList;
+import static net.andreinc.mockneat.unit.objects.From.from;
 
 import java.util.List;
+
+import org.hisp.dhis.cache.EntitiesCache;
+import org.hisp.dhis.cache.Program;
 
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
@@ -14,14 +19,14 @@ import io.restassured.specification.RequestSpecification;
 /**
  * @author Maikel Arabori <maikelarabori@gmail.com>
  */
-public class GetHeavyAnalyticsTaskNoRandom
+public class GetHeavyAnalyticsRandomTask
     extends
     DhisAbstractTask
 {
     /**
      * Configurable constants
      */
-    private static final String ORG_UNIT = "ou:ImspTQPwCqd";
+    private static final String ORG_UNIT = "ou:%s";
 
     private static final String PERIOD = "pe:LAST_12_MONTHS";
 
@@ -39,27 +44,30 @@ public class GetHeavyAnalyticsTaskNoRandom
             "n0GE1ISYrdM", "nfG18MJZX5o", "nkjlWUMIdHh", "oTOpKabJA1v", "puykO1tbcdi", "qmCKRmVj4WX", "ryFBhJdetdM",
             "sB79w2hiLp8", "sMTMkudvLCD", "tcs5YGnjiKo", "ulgL07PF8rq", "vKWOc4itBo2", "vihpFUg2WTy" ) );
 
-    private static final int API_VERSION = 30;
+    private final EntitiesCache entitiesCache;
+    private final int apiVersion;
 
-    public GetHeavyAnalyticsTaskNoRandom( final int weight )
+    public GetHeavyAnalyticsRandomTask(final int weight, final int apiVersion, final EntitiesCache entitiesCache )
     {
         this.weight = weight;
+        this.apiVersion = apiVersion;
+        this.entitiesCache = entitiesCache;
     }
 
     public String getName()
     {
-        return "(static) GET " + query();
+        return "(random) GET " + query();
     }
 
     public void execute()
     {
-        // Warm-up
-        final RequestSpecification request = given().queryParam( "filter", ORG_UNIT )
+        // Given
+        final Program aRandomProgram = randomProgram();
+        final String aRandomOrgUnitUid = randomOrgUnitUid( aRandomProgram.getOrgUnits() );
+        final RequestSpecification request = given().queryParam( "filter", format( ORG_UNIT, aRandomOrgUnitUid ) )
             .queryParam( "dimension", DIMENSIONS ).queryParam( "dimension", PERIOD )
             .queryParam( "displayProperty", DISPLAY_PROPERTY ).queryParam( "skipMeta", SKIP_META )
             .queryParam( "includeNumDen", INCLUDE_NUM_DEN );
-
-        request.get( endpoint() );
 
         // Test
         final Response response = request.get( endpoint() );
@@ -77,7 +85,7 @@ public class GetHeavyAnalyticsTaskNoRandom
 
     private String endpoint()
     {
-        return "/api/" + API_VERSION + "/analytics.json";
+        return "/api/" + apiVersion + "/analytics.json";
     }
 
     private String query()
@@ -85,5 +93,15 @@ public class GetHeavyAnalyticsTaskNoRandom
         final List<String> params = asList( ORG_UNIT, PERIOD, DIMENSIONS, DISPLAY_PROPERTY, valueOf( SKIP_META ),
             valueOf( INCLUDE_NUM_DEN ) );
         return endpoint() + " with params: " + join( " # ", params );
+    }
+
+    private String randomOrgUnitUid( final List<String> programOrgUnits )
+    {
+        return from( programOrgUnits ).get();
+    }
+
+    private Program randomProgram()
+    {
+        return from( entitiesCache.getPrograms() ).get();
     }
 }
