+++ src/main/resources/locust.properties
@@ -3,3 +3,6 @@ locust.master.port=5557
 locust.master.host=127.0.0.1
 #target.baseuri=https://play.dhis2.org/dev
 target.baseuri=http://localhost:8080/dhis
+
+# API versions
+analytics.api.version=30
